requirements.txt
# Usage: pip install -r requirements.txt
gdal
osgeo
gdal2tiles
pillow
azure-storage-blob
